+++ src/regex.ts
@@ -1,4 +1,4 @@
-import { PromptException } from "./dictionary/dictionary";
+import { PromptException, standardizeWord } from "./dictionary/dictionary";
 import { getNormalLetters, getPromptLetters } from "./emoji-renderer";
 
 /*
@@ -76,6 +76,67 @@ function trimArrows(string: string): string {
   return string.replace(/^<|>$/g, "");
 }
 
+
+/**
+ * A regular expression used to determine if a search is regex or not.
+ */
+const regexTest = /(?:^| )\/(.*)\/(?: |$)/;
+
+/**
+ * Returns a regex used for searching based on a query.
+ * The query may be in a prompt format (AB) or regex format (/AB/).
+ * 
+ * The function will interpret either format and return a regex pattern.
+ *
+ * @param promptQuery A query string to convert to regex
+ * @returns A regular expression pattern based on the query. If the query contains a valid regular expression, the function constructs and returns the regex pattern. If the query does not contain a valid regular expression, the function constructs and returns a regex pattern that matches the query as a literal string.
+ * 
+ * @example
+ * ```typescript
+ * getPromptRegexFromPromptSearch("AB") // new RegExp("AB")
+ * getPromptRegexFromPromptSearch("/A.B$/") // new RegExp("A.B$")
+ * ```
+ */
+export function getPromptRegexFromPromptSearch(promptQuery: string): RegExp {
+  let cleanQuery = standardizeWord(promptQuery);
+  let regexResult = regexTest.exec(cleanQuery);
+
+  // TODO find args in the query
+
+  // let's just be safe with backticks
+  if (cleanQuery.includes("`")) {
+    throw new PromptException("The prompt you've entered is invalid.");
+  }
+
+  if (regexResult) {
+    // This has regex
+
+    let regexInput = regexResult[1];
+
+    if (regexInput === "") {
+      // This regex is empty
+      throw new PromptException("The regex you've entered is empty.");
+    }
+
+    // check if the regex is valid
+    return validateRegex(regexInput);
+  } else {
+    // This isn't regex
+
+    // will this even work? I don't know. I'm not a regex expert. I'm just a guy who wants to make a bot. :(
+    return new RegExp(escapeRegExp(cleanQuery).replace(/\\\?|\\\./g, "."));
+  }
+}
+
+/**
+ * Escapes all RegExp special characters.
+ *
+ * @param string The input string to be escaped.
+ */
+export function escapeRegExp(string: string): string {
+  return string.replace(/[.*+?^${}()|[\]\\]/g, "\\$&"); // $& means the whole matched string
+}
+
 /**
  * Helper function to make sure a regex is valid and renames the groups in it.
  * 
@@ -91,7 +152,7 @@ export function validateRegex(regex: string): RegExp {
     r = renameRegexGroups(r);
     return r;
   } catch {
-    throw new PromptException("Invalid regex");
+    throw new PromptException("The regex you've entered is invalid.");
   }
 }
