+++ vivi/commands/count.ts
@@ -2,8 +2,8 @@ import { replyToInteraction } from '../../src/command-handler';
 import { CommandInteraction, SlashCommandBuilder } from 'discord.js';
 import { formatNumber } from '../../src/utils';
 
-import { cleanWord, solvePromptWithTimeout, getPromptRegexFromPromptSearch } from '../../src/dictionary/dictionary';
-import { getPromptRegexDisplayText } from '../../src/regex';
+import { cleanWord, solvePromptWithTimeout } from '../../src/dictionary/dictionary';
+import { getPromptRegexDisplayText, getPromptRegexFromPromptSearch } from '../../src/regex';
 
 export const data = new SlashCommandBuilder()
   .setName('count')
